Generalized the concept of commands sub groups and moved it to cmd_util.py. Currently it's used only on apio top level command but we will use it while splitting other commands to subcommands.
